edit after review

+++ src/main/java/ru/otus/dao/book/BookDao.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface BookDao {
     long count();
 
-    void insert(String name, long authorId, long genreId);
+    void insert(String name, int authorId, int genreId);
 
-    Book getById(long id);
+    Book getById(int id);
 
     List<Book> getAll();
 
-    void deleteById(long id);
+    void deleteById(int id);
 }
