+++ src/main/java/ru/otus/shell/BooksEventsCommands.java
@@ -18,13 +18,13 @@ public class BooksEventsCommands {
     }
 
     @ShellMethod(value = "Add new book", key = {"add-book"})
-    public void insertBook(String name, long authorId, long genreId) {
+    public void insertBook(String name, int authorId, int genreId) {
         bookDao.insert(name, authorId, genreId);
         System.out.printf("Книга с именем %s добавлена%n", name);
     }
 
     @ShellMethod(value = "Get book by id", key = {"get-book-by-id"})
-    public void getBookById(long id) {
+    public void getBookById(int id) {
         Book book = bookDao.getById(id);
         System.out.printf("id: %s, Название книги: %s, Имя автора: %s, Жанр: %s%n", book.getId(), book.getName(), book.getAuthor_id(), book.getGenre_id());
     }
@@ -36,7 +36,7 @@ public class BooksEventsCommands {
     }
 
     @ShellMethod(value = "Delete book by id", key = {"delete-book-by-id"})
-    public void deleteBookById(long id) {
+    public void deleteBookById(int id) {
         bookDao.deleteById(id);
         System.out.printf("Книга с id = %s удалена%n", id);
     }
